requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
defusedxml==0.7.1
dj-rest-auth==7.1.1
Django==6.0.3
django-allauth==65.14.3
django-auditlog==3.4.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
pillow==12.1.1
pipreqs==0.4.13
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
yarg==0.1.10
